pyproject.toml
[tool.poetry]
name = "yalla_business_api"
version = "0.1.0"
description = ""
authors = ["mohammedghafri <eng.m.ghafri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
psycopg2 = "^2.8.6"
djangorestframework = "^3.12.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
